ports #89811, fixes resizable textareas, and more
