+++ code/modules/asset_cache/assets/tgui.dm
@@ -2,6 +2,7 @@
 	keep_local_name = TRUE
 	assets = list(
 		"tgui.bundle.js" = file("tgui/public/tgui.bundle.js"),
+		"tgui.bundle.js.map" = file("tgui/public/tgui.bundle.js.map"),
 		"tgui.bundle.css" = file("tgui/public/tgui.bundle.css"),
 	)
 
@@ -9,5 +10,6 @@
 	keep_local_name = TRUE
 	assets = list(
 		"tgui-panel.bundle.js" = file("tgui/public/tgui-panel.bundle.js"),
+		"tgui-panel.bundle.js.map" = file("tgui/public/tgui-panel.bundle.js.map"),
 		"tgui-panel.bundle.css" = file("tgui/public/tgui-panel.bundle.css"),
 	)
