+++ tgui/babel.config.js
@@ -28,7 +28,7 @@ const createBabelConfig = (options) => {
       }],
       require.resolve('@babel/plugin-transform-jscript'),
       require.resolve('babel-plugin-inferno'),
-      removeConsole && require.resolve('babel-plugin-transform-remove-console'),
+      // removeConsole && require.resolve('babel-plugin-transform-remove-console'),
       require.resolve('common/string.babel-plugin.cjs'),
       ...plugins,
     ].filter(Boolean),
