+++ tgui/package.json
@@ -31,6 +31,7 @@
     "@types/node": "14.x",
     "@types/webpack": "^5.28.5",
     "@types/webpack-env": "^1.18.4",
+    "@types/wicg-file-system-access": "^2023.10.5",
     "@typescript-eslint/parser": "^5.62.0",
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
