+++ tgui/packages/tgui-panel/chat/middleware.js
@@ -12,6 +12,7 @@ import {
   addHighlightSetting,
   removeHighlightSetting,
   updateHighlightSetting,
+  importSettings,
 } from '../settings/actions';
 import { selectSettings } from '../settings/selectors';
 import {
@@ -96,12 +97,14 @@ export const chatMiddleware = (store) => {
   chatRenderer.events.on('scrollTrackingChanged', (scrollTracking) => {
     store.dispatch(changeScrollTracking(scrollTracking));
   });
-  setInterval(() => {
-    saveChatToStorage(store);
-  }, MESSAGE_SAVE_INTERVAL);
   return (next) => (action) => {
     const { type, payload } = action;
-    if (!initialized) {
+    const settings = selectSettings(store.getState());
+    // Load the chat once settings are loaded
+    if (!initialized && settings.initialized) {
+      setInterval(() => {
+        saveChatToStorage(store);
+      }, MESSAGE_SAVE_INTERVAL);
       initialized = true;
       loadChatFromStorage(store);
     }
@@ -175,13 +178,15 @@ export const chatMiddleware = (store) => {
       type === loadSettings.type ||
       type === addHighlightSetting.type ||
       type === removeHighlightSetting.type ||
-      type === updateHighlightSetting.type
+      type === updateHighlightSetting.type ||
+      type === updateHighlightSetting.type ||
+      type === importSettings.type
     ) {
       next(action);
-      const settings = selectSettings(store.getState());
+      const nextSettings = selectSettings(store.getState());
       chatRenderer.setHighlight(
-        settings.highlightSettings,
-        settings.highlightSettingById,
+        nextSettings.highlightSettings,
+        nextSettings.highlightSettingById,
       );
 
       return;
