+++ tgui/packages/tgui-panel/settings/settingsImExport.ts
@@ -1,0 +1,37 @@
+import { Dispatch, useDispatch } from 'common/redux';
+import type { Page } from '../chat/types';
+import { importSettings } from './actions';
+
+export function exportChatSettings(
+  settings: Record<string, any>,
+  pages: Record<string, Page>[],
+) {
+  const filename = `ss13-chatsettings-${new Date().toJSON().slice(0, 10)}.json`;
+  const mimeType = 'application/json';
+
+  const pagesEntry: Record<string, Page>[] = [];
+  pagesEntry['chatPages'] = pages;
+
+  const exportObject = Object.assign(settings, pagesEntry);
+
+  const jsonString = JSON.stringify(exportObject, null, ' ');
+
+  Byond.saveBlob(new Blob([jsonString], { type: mimeType }), filename, '.json');
+}
+
+export function importChatSettings(
+  dispatch: Dispatch,
+  settings: string | string[],
+) {
+  if (Array.isArray(settings)) {
+    return;
+  }
+  const ourImport = JSON.parse(settings);
+  if (!ourImport?.version) {
+    return;
+  }
+  const pageRecord = ourImport['chatPages'];
+  delete ourImport['chatPages'];
+
+  dispatch(importSettings(ourImport, pageRecord));
+}
