+++ tgui/packages/tgui-panel/telemetry.js
@@ -38,6 +38,14 @@ export const telemetryMiddleware = (store) => {
       Byond.sendMessage('telemetry', { connections });
       return;
     }
+    // For whatever reason we didn't get the telemetry, re-request
+    if (type === 'testTelemetryCommand') {
+      setTimeout(() => {
+        if (!telemetry) {
+          Byond.sendMessage('ready');
+        }
+      }, 500);
+    }
     // Keep telemetry up to date
     if (type === 'backend/update') {
       next(action);
