+++ tgui/packages/tgui/styles/components/TextArea.scss
@@ -44,6 +44,7 @@ $border-radius: Input.$border-radius !default;
 .TextArea__textarea {
   display: block;
   position: absolute;
+  resize: none;
   top: 0;
   bottom: 0;
   left: 0;
